refactor opponent card rendering
