+++ src/helpers/classicBattle.js
@@ -2,7 +2,7 @@ export * from "./classicBattle/roundManager.js";
 export * from "./classicBattle/selectionHandler.js";
 export * from "./classicBattle/quitModal.js";
 export {
-  revealOpponentCard,
+  renderOpponentCard,
   enableNextRoundButton,
   disableNextRoundButton,
   updateDebugPanel
@@ -10,3 +10,4 @@ export {
 export { getOpponentJudoka } from "./classicBattle/cardSelection.js";
 export { scheduleNextRound } from "./classicBattle/timerService.js";
 export { applyRoundUI } from "./classicBattle/roundUI.js";
+export { getOpponentCardData } from "./classicBattle/opponentController.js";
